Moves reference star model into models module
