+++ tests/test_models/test_supernovae/testObservedCadence.py
