+++ tests/test_pipeline/testFittingPipeline.py
